requirements-gcp.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=gcp --output-file=requirements-gcp.txt
#
asyncstdlib==3.10.5
    # via ewms-pilot (setup.py)
cachetools==5.2.0
    # via google-auth
certifi==2022.12.7
    # via
    #   pulsar-client
    #   requests
charset-normalizer==2.1.1
    # via requests
google-api-core[grpc]==2.11.0
    # via
    #   google-cloud-pubsub
    #   wipac-mqclient
google-auth==2.15.0
    # via
    #   google-api-core
    #   wipac-mqclient
google-cloud-pubsub==2.13.11
    # via wipac-mqclient
googleapis-common-protos[grpc]==1.57.1
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
    #   wipac-mqclient
grpc-google-iam-v1==0.12.4
    # via google-cloud-pubsub
grpcio==1.51.1
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.51.1
    # via
    #   google-api-core
    #   google-cloud-pubsub
idna==3.4
    # via requests
proto-plus==1.22.1
    # via google-cloud-pubsub
protobuf==4.21.12
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pulsar-client==3.0.0
    # via wipac-mqclient
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
requests==2.28.1
    # via
    #   google-api-core
    #   wipac-dev-tools
rsa==4.9
    # via google-auth
six==1.16.0
    # via google-auth
typing-extensions==4.4.0
    # via wipac-dev-tools
urllib3==1.26.13
    # via requests
wipac-dev-tools==1.6.10
    # via wipac-mqclient
wipac-mqclient[gcp,pulsar]==0.6.8
    # via ewms-pilot (setup.py)
